requirements.txt
asgiref==3.3.2
Django==3.1.8
djangorestframework==3.12.4
nose==1.3.7
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
